# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FundiVerify — Nairobi Contractor Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /contractors:
    get:
      operationId: listContractors
      tags: [contractors]
      summary: List and search verified contractors
      parameters:
        - name: trade
          in: query
          schema:
            type: string
        - name: location
          in: query
          schema:
            type: string
        - name: tier
          in: query
          schema:
            type: string
            enum: [free, pro]
        - name: minRating
          in: query
          schema:
            type: number
        - name: verified
          in: query
          schema:
            type: boolean
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of contractors
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractorListResponse"
    post:
      operationId: createContractor
      tags: [contractors]
      summary: Apply as a contractor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContractorBody"
      responses:
        "201":
          description: Contractor application submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contractor"
...
  /contractors/{id}:
    get:
      operationId: getContractor
      tags: [contractors]
      summary: Get contractor profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contractor profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractorDetail"
        "404":
          description: Not found
    patch:
      operationId: updateContractor
      tags: [contractors]
      summary: Update contractor profile or verification status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateContractorBody"
      responses:
        "200":
          description: Updated contractor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contractor"

  # ── Jobs ──
...
  /disputes:
    get:
      operationId: listDisputes
      tags: [disputes]
      summary: List disputes (admin view)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [open, under_review, resolved, closed]
      responses:
        "200":
          description: List of disputes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Dispute"
    post:
      operationId: createDispute
      tags: [disputes]
      summary: Raise a dispute for a job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDisputeBody"
      responses:
        "201":
          description: Dispute raised
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dispute"
  /disputes/{id}:
    patch:
      operationId: updateDispute
      tags: [disputes]
      summary: Update dispute status/resolution (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDisputeBody"
      responses:
        "200":
          description: Updated dispute
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dispute"

  # ── Saved Contractors ──
...
  /dashboard/contractor/{contractorId}:
    get:
      operationId: getContractorDashboard
      tags: [dashboard]
      summary: Contractor earnings and performance dashboard
      parameters:
        - name: contractorId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contractor dashboard data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractorDashboard"
components:
  schemas:
...
    Contractor:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        trade:
          type: string
        location:
          type: string
        bio:
          type: string
        avatarUrl:
          type: string
        verificationStatus:
          type: string
          enum: [pending, verified, rejected]
        subscriptionTier:
          type: string
          enum: [free, pro]
        rating:
          type: number
        reviewCount:
          type: integer
        jobsCompleted:
          type: integer
        yearsExperience:
          type: integer
        phone:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, name, trade, location, verificationStatus, subscriptionTier, rating, reviewCount, jobsCompleted, createdAt]
    ContractorDetail:
      allOf:
        - $ref: "#/components/schemas/Contractor"
        - type: object
          properties:
            portfolioPhotos:
              type: array
              items:
                type: string
            specializations:
              type: array
              items:
                type: string
            reviews:
              type: array
              items:
                $ref: "#/components/schemas/Review"
            isSaved:
              type: boolean
    ContractorListResponse:
      type: object
      properties:
        contractors:
          type: array
          items:
            $ref: "#/components/schemas/Contractor"
        total:
          type: integer
      required: [contractors, total]
...
    CreateContractorBody:
      type: object
      properties:
        name:
          type: string
        trade:
          type: string
        location:
          type: string
        bio:
          type: string
        phone:
          type: string
        yearsExperience:
          type: integer
        avatarUrl:
          type: string
        portfolioPhotos:
          type: array
          items:
            type: string
        specializations:
          type: array
          items:
            type: string
      required: [name, trade, location, phone]
    UpdateContractorBody:
      type: object
      properties:
        bio:
          type: string
        location:
          type: string
        verificationStatus:
          type: string
          enum: [pending, verified, rejected]
        subscriptionTier:
          type: string
          enum: [free, pro]
        avatarUrl:
          type: string
        portfolioPhotos:
          type: array
          items:
            type: string
        yearsExperience:
          type: integer
    Job:
      type: object
      properties:
        id:
          type: integer
        homeownerId:
          type: integer
        homeownerName:
          type: string
        title:
          type: string
        trade:
          type: string
        location:
          type: string
        description:
          type: string
        estimatedBudget:
          type: number
        status:
          type: string
          enum: [open, quoted, in_progress, completed, disputed, cancelled]
        quoteCount:
          type: integer
        urgency:
          type: string
          enum: [flexible, within_week, asap]
        createdAt:
          type: string
          format: date-time
      required: [id, homeownerId, title, trade, location, description, status, quoteCount, urgency, createdAt]
...
    Review:
      type: object
      properties:
        id:
          type: integer
        jobId:
          type: integer
        contractorId:
          type: integer
        homeownerId:
          type: integer
        homeownerName:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
        jobTitle:
          type: string
        jobTrade:
          type: string
        photos:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [pending, approved, rejected]
        createdAt:
          type: string
          format: date-time
      required: [id, jobId, contractorId, homeownerId, homeownerName, rating, comment, status, createdAt]
...
    Dispute:
      type: object
      properties:
        id:
          type: integer
        jobId:
          type: integer
        jobTitle:
          type: string
        raisedById:
          type: integer
        raisedByName:
          type: string
        reason:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [open, under_review, resolved, closed]
        resolution:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, jobId, raisedById, raisedByName, reason, status, createdAt]
    CreateDisputeBody:
      type: object
      properties:
        jobId:
          type: integer
        raisedById:
          type: integer
        raisedByName:
          type: string
        reason:
          type: string
        description:
          type: string
      required: [jobId, raisedById, raisedByName, reason]
    UpdateDisputeBody:
      type: object
      properties:
        status:
          type: string
          enum: [open, under_review, resolved, closed]
        resolution:
          type: string
...
    ContractorDashboard:
      type: object
      properties:
        contractorId:
          type: integer
        totalEarnings:
          type: number
        thisMonthEarnings:
          type: number
        jobsCompleted:
          type: integer
        activeJobs:
          type: integer
        pendingQuotes:
          type: integer
        averageRating:
          type: number
        profileViews:
          type: integer
        leadsReceived:
          type: integer
        earningsByMonth:
          type: array
          items:
            type: object
            properties:
              month:
                type: string
              amount:
                type: number
            required: [month, amount]
        recentJobs:
          type: array
          items:
            $ref: "#/components/schemas/Job"
      required: [contractorId, totalEarnings, thisMonthEarnings, jobsCompleted, activeJobs, pendingQuotes, averageRating, profileViews, leadsReceived, earningsByMonth, recentJobs]