# lib/api-zod/src/generated/types/contractor.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FundiVerify — Nairobi Contractor Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ContractorSubscriptionTier } from "./contractorSubscriptionTier";
import type { ContractorVerificationStatus } from "./contractorVerificationStatus";

export interface Contractor {
  id: number;
  name: string;
  trade: string;
  location: string;
  bio?: string;
  avatarUrl?: string;
  verificationStatus: ContractorVerificationStatus;
  subscriptionTier: ContractorSubscriptionTier;
  rating: number;
  reviewCount: number;
  jobsCompleted: number;
  yearsExperience?: number;
  phone?: string;
  createdAt: Date;
}