# lib/api-zod/src/generated/types/updateContractorBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FundiVerify — Nairobi Contractor Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateContractorBodySubscriptionTier } from "./updateContractorBodySubscriptionTier";
import type { UpdateContractorBodyVerificationStatus } from "./updateContractorBodyVerificationStatus";

export interface UpdateContractorBody {
  bio?: string;
  location?: string;
  verificationStatus?: UpdateContractorBodyVerificationStatus;
  subscriptionTier?: UpdateContractorBodySubscriptionTier;
  avatarUrl?: string;
  portfolioPhotos?: string[];
  yearsExperience?: number;
}